attendance-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose==3.3.0
pydantic==2.5.0
pymysql==1.1.0
asyncpg==0.29.0
redis==5.0.1
python-multipart==0.0.6
opencv-python==4.8.1
numpy==1.26.0
onnxruntime==1.16.0
uvloop==0.17.0
